# Optimize StaticSparseDAG