@messense 2020-11-27
Reverted, it actually makes the weicheng example a lot slower.